Changed Tides name, Url and Description
Tides 2014 URL is pointing to the Tides 2018 version, I've changed it to point to the predecessor. I took a guess at the name and called it 'Tides (2014)', changed the URL to point to the older version and changed the description for accuracy.

While I was in there, I changed domain names to HTTPS; simple stuff, might not be accurate to what you want, but figured I'd contribute.